Merged the n_dims and all the sparse atoms stuff

+++ wepy/reporter/reporter.py
@@ -59,7 +59,8 @@ class WalkersPickleReporter(Reporter):
                 pickle.dump(walkers, wf)
 
             # remove old pickles if we have more than the `num_backups`
-            if (cycle_idx // self.backup_freq) >= self.num_backups:
-                old_idx = cycle_idx - self.num_backups * self.backup_freq
-                old_pkl_fname = "walkers_cycle_{}.pkl".format(old_idx)
-                os.remove(osp.join(self.save_dir, old_pkl_fname))
+            if self.num_backups is not None:
+                if (cycle_idx // self.backup_freq) >= self.num_backups:
+                    old_idx = cycle_idx - self.num_backups * self.backup_freq
+                    old_pkl_fname = "walkers_cycle_{}.pkl".format(old_idx)
+                    os.remove(osp.join(self.save_dir, old_pkl_fname))
